+++ src/main/groovy/com/trevorism/data/provider/PassThruSecureHttpClient.groovy
@@ -1,0 +1,11 @@
+package com.trevorism.data.provider
+
+import com.trevorism.https.SecureHttpClient
+import com.trevorism.https.SecureHttpClientBase
+
+@jakarta.inject.Singleton
+class PassThruSecureHttpClient extends SecureHttpClientBase implements SecureHttpClient{
+    PassThruSecureHttpClient(PassThruObtainTokenStrategy strategy) {
+        super(strategy)
+    }
+}